[linker] Replace RemoveCode with xml-based substitutions
The core of the step was to remove code that is not part of net5.

The extra iOS/Mac specific steps can use substitutions. The main
difference is that the method (metadata) and calls (IL) are not
removed, i.e. the method becomes just an empty stub, so it has less
impact (than before) on the application size (or performance).
Since those are not critical (in number of sites or calls) then it
likely does not matter much.