[msbuild] Fix linking after build changes
https://github.com/xamarin/xamarin-macios/commit/973ecf4afad0be4f4a73b3d94eff19a7ad0b349d
changed the order of operations quite a bit and regressed the linker

First `<IsTrimmable>true</IsTrimmable>` was not applied anymore on
`Xamarin.iOS.dll` so it went from 200kb to 18MB (easy to notice).
Turns out `ResolvedFileToPublish` was still empty when
`_SelectAssembliesToLink` called. Adding a dependency on
`ComputeResolvedFilesToPublishList` solved this.

Next `registerentryassembly.xml` was applied as a substitution - but
that's fine for non simulator builds only. Moving the `ItemGroup` into
`_ComputeLinkerSubstitutions` solved it (I think it was computed too
early).