[linker] Enable RemoveAttributes step
Size impact varies quite a bit depending on the assemblies
https://gist.github.com/spouliot/2310107afd8649f1268287023b84d70d

The list of attributes will need to be updated for dotnet SDK. Ideally
most of the dotnet specific attributes will be handled upstream.

We still need to keep the old BCL ones for the SDK we ship for pre-net5
builds... but we'll be able to disable them.