[linker] Semi-enable MetadataReducerSubStep
`ReflectedParameterNames` cannot be automagically detected anymore since:

* Existing code requires sub-classing `MarkStep` for detecting if user code is reflecting; and
* We don't ship the BCL as part of the SDK, so things might changes without our knowledge

It's already possible (old `mtouch`) to opt-out the optimization.
An alternative is to make it opt-in or get more assurance from the SDK or the new linker reflection detection.

**Impact**

* [6% smaller managed code (simulator) when removing parameter names](https://gist.github.com/spouliot/e4ad9345bbc677e1920c9486b3965a26)

Some assemblies shows 0% because they are block based (512 bytes) and the difference was under that limit.
A lot of small assemblies makes the _rounding_ more apparent.

* [0% without removing parameter names](https://gist.github.com/spouliot/b6906450a14b2d74d6b6052ae6c2dc73)

Xamarin.iOS.dll is 0.3% smaller by sharing p/invoke names, but not enough of them are around to make a significant difference.