[linker] Add temporary ExtraMarkStep and ExtraSweepStep
Mostly related to https://github.com/mono/linker/issues/1188
This should be removed once everything is available upstream.

Part of this was merged recently with https://github.com/mono/linker/pull/1193
but the current dotnet package we use does not have the bits.

Code makes it easy to confirm mono/linker code does the same thing
as we expect. Nothing else should be marked, so a breakpoint on the
`Mark` calls (when DEBUG is defined) should never be hit.